+++ client/Makefile.am
@@ -20,11 +20,32 @@ bin_PROGRAMS = snapper
 
 snapper_SOURCES =					\
 	snapper.cc					\
+	cmd.h						\
+	cmd-list-configs.cc				\
+	cmd-create-config.cc				\
+	cmd-delete-config.cc				\
+	cmd-get-config.cc				\
+	cmd-set-config.cc				\
+	cmd-list.cc					\
+	cmd-create.cc					\
+	cmd-modify.cc					\
+	cmd-delete.cc					\
+	cmd-mount.cc					\
+	cmd-umount.cc					\
+	cmd-status.cc					\
+	cmd-diff.cc					\
+	cmd-xadiff.cc					\
+	cmd-undochange.cc				\
+	cmd-rollback.cc					\
+	cmd-setup-quota.cc				\
+	cmd-cleanup.cc					\
+	cmd-debug.cc					\
 	cleanup.cc		cleanup.h		\
 	proxy.cc		proxy.h			\
 	proxy-dbus.cc		proxy-dbus.h		\
 	proxy-lib.cc		proxy-lib.h		\
 	misc.cc			misc.h			\
+	MyFiles.cc		MyFiles.h		\
 	Options.cc		Options.h		\
 	GlobalOptions.cc	GlobalOptions.h		\
 	Command.cc		Command.h
